Restructure library's folder structure
