+++ examples/normals_estimation.cpp
@@ -2,10 +2,10 @@
 #include <Eigen/Eigenvalues>
 #include <iostream>
 #include <pcp/io/ply.hpp>
-#include <pcp/normal.hpp>
-#include <pcp/octree.hpp>
-#include <pcp/point.hpp>
-#include <pcp/point_view.hpp>
+#include <pcp/common/normal.hpp>
+#include <pcp/octree/octree.hpp>
+#include <pcp/common/point.hpp>
+#include <pcp/common/point_view.hpp>
 
 struct vertex_t
 {
