--- include/pcp.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#include "pcp/intersections.hpp"
-#include "pcp/io/endianness.hpp"
-#include "pcp/io/obj.hpp"
-#include "pcp/io/ply.hpp"
-#include "pcp/normal.hpp"
-#include "pcp/octree.hpp"
-#include "pcp/point.hpp"
