+++ include/pcp/pcp.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "common/intersections.hpp"
+#include "common/normal.hpp"
+#include "common/octree.hpp"
+#include "common/point.hpp"
+#include "io/endianness.hpp"
+#include "io/obj.hpp"
+#include "io/ply.hpp"
